invo/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0  # PostgreSQL async driver
alembic==1.12.1  # Database migrations

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI endpoints

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Observability & Monitoring
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
prometheus-client==0.19.0
structlog==23.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid-utils==0.7.0

# Development
pre-commit==3.6.0

# Cloud & Backup
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Performance & Load Testing  
locust==2.17.0